Refactor shading
Remove hittable random point as this is Cook-style and not by Whitted-Style
Move scatter logic to renderer, material is in charge of filling in part of
ray payload. Logic is now in switch statement which is less object-oriented
and more data oriented.
Merge lights and geometry to trace with same call. Probability of hitting
a point light is very low for a random ray (zero in continous space).
Lights are still accessible with indices which are used to spawn a shadow
ray per light.